pyproject.toml
[project]
name = "ag-ui-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "TheGreatBonnie",email = "mwendabkaberia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "ag-ui-protocol (>=0.1.5,<0.2.0)",
    "openai (>=1.88.0,<2.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "langchain-core (>=0.3.65,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
